requirements.txt
dbt-dremio
astronomer-cosmos